#209 fix latest work requests view

+++ service_packs/2016_06_29_1_edit_v_LatestWorkRequests.sql
@@ -1,4 +1,4 @@
-﻿IF OBJECT_ID ('dbo.v_LatestWorkRequests', N'V') IS NOT NULL
+IF OBJECT_ID ('dbo.v_LatestWorkRequests', N'V') IS NOT NULL
    DROP VIEW dbo.v_LatestWorkRequests;
 GO
